Revert "[nrf noup] [openthread] Fixed wrong casting"
This reverts commit 0beaf1fa21869f6c713d18c60891652022fe8269.

+++ src/platform/OpenThread/GenericThreadStackManagerImpl_OpenThread.cpp
@@ -1869,7 +1869,7 @@ GenericThreadStackManagerImpl_OpenThread<ImplClass>::SetSEDIntervalMode(Connecti
     {
 #if CHIP_DEVICE_CONFIG_THREAD_SSED
         // Set CSL period in units of 10 symbols, convert it to microseconds and divide by 1000 to get milliseconds.
-        otErr         = otLinkCslSetPeriod(mOTInst, static_cast<uint16_t>(static_cast<uint32_t>(interval.count() * 1000) / static_cast<uint32_t>(OT_US_PER_TEN_SYMBOLS)));
+        otErr         = otLinkCslSetPeriod(mOTInst, interval.count() * 1000 / OT_US_PER_TEN_SYMBOLS);
         curIntervalMS = otLinkCslGetPeriod(mOTInst) * OT_US_PER_TEN_SYMBOLS / 1000;
 #else
         otErr         = otLinkSetPollPeriod(mOTInst, interval.count());
